fix: Fix the Monaco error from showing up (#33432)
Fix the Monaco error from showing up.
